requirements.txt
ExifRead==2.1.2
Pillow==6.1.0
Jinja2==2.10.1